# docs/api/mvp-booking-auth-messages-openapi.yaml
openapi: 3.0.3
info:
  title: Aviaframe MVP - Booking/Auth/Messages API
  version: 2026-02-26
servers:
  - url: https://admin.aviaframe.com/api
paths:
...
  /auth/claim-order:
    post:
      summary: Link anonymous order to authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claim_token]
              properties:
                claim_token:
                  type: string
      responses:
        '200':
          description: Claimed
        '404':
          description: Claim token not found
        '410':
          description: Claim token expired
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT